# server/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.RackAssignRequest:
    properties:
      id_rack:
        example: 2
        type: integer
    type: object
...
  handlers.SpoolCreateRequest:
    properties:
      id_filament:
        example: uuid-here
        type: string
      id_rack:
        example: 3
        type: integer
      is_dry:
        example: false
        type: boolean
      is_spooled:
        example: true
        type: boolean
      notes:
        example: opened 2024-01
        type: string
      weight_remaining_grams:
        example: 850
        type: number
    type: object
  handlers.SpoolUpdateRequest:
    properties:
      id_rack:
        type: integer
      is_dry:
        example: false
        type: boolean
      is_spooled:
        example: true
        type: boolean
      notes:
        type: string
      weight_remaining_grams:
        example: 650
        type: number
    type: object
  handlers.TagRequest:
    properties:
      color:
        example: '#FF5733'
        type: string
      icon:
        type: string
      name:
        example: Urgent
        type: string
    type: object
  handlers.WeightUpdateRequest:
    properties:
      weight_remaining_grams:
        example: 520
        type: number
    type: object
...
  models.Brand:
    properties:
      created_at:
        type: string
      empty_spool_weight_grams:
        type: number
      id_brand:
        type: string
      id_origin:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
  models.Filament:
    properties:
      bed_temp_max:
        type: integer
      bed_temp_min:
        type: integer
      brand:
        $ref: '#/definitions/models.Brand'
      color_hex:
        type: string
      color_info:
        type: object
      created_at:
        type: string
      density:
        type: number
      dry_temp:
        type: integer
      dry_time:
        type: integer
      filled_type:
        type: string
      id_brand:
        type: string
      id_filament:
        type: string
      id_origin:
        type: string
      image_url:
        type: string
      material_type:
        type: string
      nozzle_temp_max:
        type: integer
      nozzle_temp_min:
        type: integer
      product_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      weight_grams:
        type: integer
    type: object
...
  models.RackRef:
    properties:
      id_rack:
        type: integer
      name:
        type: string
    type: object
  models.Spool:
    properties:
      acquired_at:
        type: string
      filament:
        $ref: '#/definitions/models.Filament'
      id_filament:
        type: string
      id_rack:
        type: integer
      id_spool:
        type: integer
      id_user:
        type: string
      is_dry:
        type: boolean
      is_spooled:
        type: boolean
      notes:
        type: string
      rack:
        $ref: '#/definitions/models.RackRef'
      weight_remaining_grams:
        type: number
    type: object
  models.Tag:
    properties:
      color:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id_tag:
        type: integer
      id_user:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Filament inventory management for 3D printing enthusiasts.
  title: Filaventory API
  version: "1.0"
paths:
...
  /api/spools:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by material type
        in: query
        name: material
        type: string
      - description: Filter by rack ID ('none' for unracked)
        in: query
        name: rack_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Spool'
            type: array
      security:
      - BearerAuth: []
      summary: List spools
      tags:
      - spools
    post:
      consumes:
      - application/json
      parameters:
      - description: Spool data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SpoolCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Spool'
        "400":
          description: id_filament is required
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a spool
      tags:
      - spools
  /api/spools/{id}:
    delete:
      parameters:
      - description: Spool ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a spool
      tags:
      - spools
    get:
      parameters:
      - description: Spool ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spool'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a spool
      tags:
      - spools
    put:
      consumes:
      - application/json
      parameters:
      - description: Spool ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SpoolUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spool'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a spool
      tags:
      - spools
  /api/spools/{id}/rack:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Spool ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rack ID (null to unassign)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RackAssignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spool'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Assign or unassign a spool from a rack
      tags:
      - spools
  /api/spools/{id}/weight:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Spool ID
        in: path
        name: id
        required: true
        type: integer
      - description: New weight
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WeightUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spool'
        "400":
          description: weight must be > 0
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update remaining weight of a spool
      tags:
      - spools
  /api/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tag'
            type: array
      security:
      - BearerAuth: []
      summary: List tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      parameters:
      - description: Tag data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tag'
        "400":
          description: name is required
          schema:
            type: string
        "409":
          description: tag already exists
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a tag
      tags:
      - tags
  /api/tags/{id}:
    delete:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a tag
      tags:
      - tags
    put:
      consumes:
      - application/json
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tag'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a tag
      tags:
      - tags
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" in the Authorization header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"